add main screen view model

+++ WeatherForecast/app/src/main/java/com/nasyxnadeem/weatherforecast/model/Coord.kt
@@ -1,6 +1,6 @@
 package com.nasyxnadeem.weatherforecast.model
 
 data class Coord(
-    val lat: Int,
-    val lon: Int
+    val lat: Double,
+    val lon: Double
 )
